Fixed some tests and moved some tests around
