+++ spaceship/spec/connect_api/models/app_spec.rb
