+++ spaceship/spec/connect_api/models/beta_app_localizations_spec.rb
