+++ spaceship/spec/connect_api/models/beta_app_review_detail_spec.rb
