+++ spaceship/spec/connect_api/models/beta_build_metric_spec.rb
