+++ spaceship/spec/connect_api/models/beta_group_spec.rb
