+++ spaceship/spec/connect_api/models/beta_tester_metric_spec.rb
