+++ spaceship/spec/connect_api/models/beta_tester_spec.rb
