+++ spaceship/spec/connect_api/models/build_beta_detail_spec.rb
