+++ spaceship/spec/connect_api/models/build_spec.rb
