+++ spaceship/spec/connect_api/models/bundle_id_spec.rb
