+++ spaceship/spec/connect_api/models/user_spec.rb
