+++ spaceship/spec/connect_api/users/user_client_spec.rb
@@ -1,0 +1,63 @@
+describe Spaceship::ConnectAPI::Users::Client do
+  let(:client) { Spaceship::ConnectAPI::Users::Client.instance }
+  let(:hostname) { Spaceship::ConnectAPI::Users::Client.hostname }
+  let(:username) { 'spaceship@krausefx.com' }
+  let(:password) { 'so_secret' }
+
+  before do
+    Spaceship::Tunes.login(username, password)
+  end
+
+  context 'sends api request' do
+    before(:each) do
+      allow(client).to receive(:handle_response)
+      module Spaceship
+        class ConnectAPI
+          module Users
+            class Client
+              include Spaceship::ConnectAPI::Users
+            end
+          end
+        end
+      end
+    end
+
+    def test_request_params(url, params)
+      req_mock = double
+      options_mock = double
+
+      allow(req_mock).to receive(:headers).and_return({})
+
+      expect(req_mock).to receive(:url).with(url)
+      expect(req_mock).to receive(:params=).with(params)
+      expect(req_mock).to receive(:options).and_return(options_mock)
+      expect(options_mock).to receive(:params_encoder=).with(Faraday::NestedParamsEncoder)
+
+      return req_mock
+    end
+
+    def test_request_body(url, body)
+      req_mock = double
+      header_mock = double
+      expect(req_mock).to receive(:url).with(url)
+      expect(req_mock).to receive(:body=).with(JSON.generate(body))
+      expect(req_mock).to receive(:headers).and_return(header_mock)
+      expect(header_mock).to receive(:[]=).with("Content-Type", "application/json")
+
+      return req_mock
+    end
+
+    describe "users" do
+      context 'get_users' do
+        let(:path) { "users" }
+
+        it 'succeeds' do
+          params = {}
+          req_mock = test_request_params(path, params)
+          expect(client).to receive(:request).with(:get).and_yield(req_mock)
+          client.get_users
+        end
+      end
+    end
+  end
+end
